+++ app/src/main/java/eu/kanade/presentation/manga/components/MarkdownRender.kt
@@ -26,7 +26,9 @@ import com.mikepenz.markdown.m3.markdownTypography
 import com.mikepenz.markdown.model.MarkdownAnnotator
 import com.mikepenz.markdown.model.markdownAnnotator
 import com.mikepenz.markdown.model.markdownPadding
+import com.mikepenz.markdown.model.rememberMarkdownState
 import org.intellij.markdown.MarkdownTokenTypes.Companion.HTML_TAG
+import org.intellij.markdown.flavours.MarkdownFlavourDescriptor
 import org.intellij.markdown.flavours.commonmark.CommonMarkFlavourDescriptor
 import org.intellij.markdown.flavours.commonmark.CommonMarkMarkerProcessor
 import org.intellij.markdown.flavours.gfm.table.GitHubTableMarkerProvider
@@ -49,12 +51,18 @@ import tachiyomi.presentation.core.components.material.padding
 fun MarkdownRender(
     content: String,
     modifier: Modifier = Modifier,
+    flavour: MarkdownFlavourDescriptor = SimpleMarkdownFlavourDescriptor,
     annotator: MarkdownAnnotator = markdownAnnotator(),
 ) {
-    Markdown(
+    val markdownState = rememberMarkdownState(
         content = content,
+        flavour = flavour,
+        immediate = true,
+    )
+
+    Markdown(
+        markdownState = markdownState,
         annotator = annotator,
-        flavour = SimpleMarkdownFlavourDescriptor,
         typography = mihonMarkdownTypography(),
         padding = mihonMarkdownPadding(),
         components = mihonMarkdownComponents(),
@@ -114,10 +122,16 @@ private fun mihonMarkdownComponents() = markdownComponents(
         val markers = listOf("•", "◦", "▸", "▹")
 
         CompositionLocalProvider(
-            LocalBulletListHandler provides { _, _, _, _ -> "${markers[ul.listDepth % markers.size]} " },
+            LocalBulletListHandler provides { _, _, _, _, _ -> "${markers[ul.listDepth % markers.size]} " },
         ) {
             Column(modifier = Modifier.padding(start = MaterialTheme.padding.small)) {
-                MarkdownBulletList(ul.content, ul.node, style = ul.typography.bullet)
+                MarkdownBulletList(
+                    content = ul.content,
+                    node = ul.node,
+                    style = ul.typography.bullet,
+                    markerModifier = { Modifier.alignBy(FirstBaseline) },
+                    listModifier = { Modifier.alignBy(FirstBaseline) },
+                )
             }
         }
     },